requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
python-multipart==0.0.6
easyocr==1.7.0
face-recognition==1.3.0
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
opencv-python-headless==4.8.0.76
typing-extensions==4.7.1
pydantic==2.3.0
starlette==0.27.0 